feat: Add user role assignment logic to createUser function
This commit adds logic to the `createUser` function in the `userController.js` file to assign the user role based on the provided role parameter. If the role is "admin", the user role will be set to "admin", otherwise it will be set to "user". This logic is added as an example and may not be suitable for all projects.